+++ source/Energinet.DataHub.MessageHub.Model.Tests/Peek/ResponseBundleParserTests.cs
@@ -148,11 +148,9 @@ namespace Energinet.DataHub.MessageHub.Model.Tests.Peek
             var valid = new DataBundleResponseDto(
                 Guid.NewGuid(),
                 "A052186D-89E1-4975-8811-2B4E6137491A",
-                new DataBundleResponseErrorDto
-                {
-                    FailureDescription = "error",
-                    Reason = DataBundleResponseErrorReason.DatasetNotAvailable
-                });
+                new DataBundleResponseErrorDto(
+                    DataBundleResponseErrorReason.DatasetNotAvailable,
+                    "error"));
 
             // act
             var actual = target.Parse(valid);