+++ source/Energinet.DataHub.PostOffice.EntryPoint.Operations/HealthCheck/ServiceBusQueueVerifier.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using Azure.Messaging.ServiceBus;
 
@@ -20,13 +21,16 @@ namespace Energinet.DataHub.PostOffice.EntryPoint.Operations.HealthCheck
 {
     public sealed class ServiceBusQueueVerifier : IServiceBusQueueVerifier
     {
+        [SuppressMessage("Reliability", "CA2007:Consider calling ConfigureAwait on the awaited task", Justification = "Issue: https://github.com/dotnet/roslyn-analyzers/issues/5712")]
         public async Task<bool> VerifyAsync(string connectionString, string name)
         {
             try
             {
                 await using var client = new ServiceBusClient(connectionString);
+
                 await using var receiver = client.CreateReceiver(name);
                 await receiver.PeekMessagesAsync(1).ConfigureAwait(false);
+
                 return true;
             }
 #pragma warning disable CA1031