+++ source/Energinet.DataHub.MessageHub.Model/documents/release-notes/release-notes.md
@@ -1,5 +1,13 @@
 # Energinet.DataHub.MessageHub.Model Release notes
 
+## Version 2.1.0
+
+- Failure description must not be null.
+
+## Version 2.0.0
+
+- .NET 6 upgrade.
+
 ## Version 1.0.0
 
 - Preparing package for release.